+++ packages/qiskit-sim/lib/gates.js
@@ -25,7 +25,8 @@ const identityMatrix = n => {
   return matrix;
 };
 
-const buildControlled = U => {
+const buildControlled = Gate => {
+  const U = Gate.matrix;
   const m = U.length;
   const C = identityMatrix(m * 2);
 
@@ -38,49 +39,67 @@ const buildControlled = U => {
   return C;
 };
 
-const gates = {
-  x: [[0, 1], [1, 0]],
-  y: [[0, math.multiply(-1, math.i)], [math.i, 0]],
-  z: [[1, 0], [0, -1]],
-  h: [
-    [1 / math.sqrt(2), 1 / math.sqrt(2)],
-    [1 / math.sqrt(2), 0 - 1 / math.sqrt(2)],
-  ],
-  srn: [
-    [1 / math.sqrt(2), 0 - 1 / math.sqrt(2)],
-    [1 / math.sqrt(2), 1 / math.sqrt(2)],
-  ],
-  s: [[1, 0], [0, math.pow(math.e, math.multiply(math.i, math.PI / 2))]],
-  r2: [[1, 0], [0, math.pow(math.e, math.multiply(math.i, math.PI / 2))]],
-  r4: [[1, 0], [0, math.pow(math.e, math.multiply(math.i, math.PI / 4))]],
-  r8: [[1, 0], [0, math.pow(math.e, math.multiply(math.i, math.PI / 8))]],
-  swap: [[1, 0, 0, 0], [0, 0, 1, 0], [0, 1, 0, 0], [0, 0, 0, 1]],
-  srswap: [
-    [1, 0, 0, 0],
-    [
-      0,
-      math.multiply(0.5, math.add(1, math.i)),
-      math.multiply(0.5, math.subtract(1, math.i)),
-      0,
-    ],
-    [
-      0,
-      math.multiply(0.5, math.subtract(1, math.i)),
-      math.multiply(0.5, math.add(1, math.i)),
-      0,
-    ],
-    [0, 0, 0, 1],
-  ],
+class Gate {
+  constructor(name, matrix) {
+    this.name = name;
+    this.matrix = matrix;
+  }
+}
+
+Gate.x = new Gate('x', [[0, 1], [1, 0]]);
+Gate.y = new Gate('y', [[0, math.multiply(-1, math.i)], [math.i, 0]]);
+Gate.z = new Gate('z', [[1, 0], [0, -1]]);
+Gate.id = new Gate('id', [[1, 0], [0, 1]]);
+Gate.h = new Gate('h',
+                  [[1 / math.sqrt(2), 1 / math.sqrt(2)],
+                   [1 / math.sqrt(2), 0 - 1 / math.sqrt(2)]]);
+Gate.srn = new Gate('srn',
+                    [[1 / math.sqrt(2), 0 - 1 / math.sqrt(2)],
+                     [1 / math.sqrt(2), 1 / math.sqrt(2)]]);
+Gate.s = new Gate('s',
+                  [[1, 0],
+                   [0, math.pow(math.e, math.multiply(math.i, math.PI / 2))]]);
+Gate.r2 = new Gate('r2',
+                   [[1, 0],
+                    [0, math.pow(math.e, math.multiply(math.i, math.PI / 2))]]);
+Gate.r4 = new Gate('r4',
+                   [[1, 0],
+                    [0, math.pow(math.e, math.multiply(math.i, math.PI / 4))]]);
+Gate.r8 = new Gate('r8',
+                   [[1, 0],
+                    [0, math.pow(math.e, math.multiply(math.i, math.PI / 8))]]);
+Gate.swap = new Gate('swap',
+                     [[1, 0, 0, 0], [0, 0, 1, 0], [0, 1, 0, 0], [0, 0, 0, 1]]);
+Gate.srswap = new Gate('srswap',
+                      [
+                        [1, 0, 0, 0],
+                        [
+                          0,
+                          math.multiply(0.5, math.add(1, math.i)),
+                          math.multiply(0.5, math.subtract(1, math.i)),
+                          0,
+                        ],
+                        [
+                          0,
+                          math.multiply(0.5, math.subtract(1, math.i)),
+                          math.multiply(0.5, math.add(1, math.i)),
+                          0,
+                        ],
+                        [0, 0, 0, 1],
+                      ]);
+Gate.cx = new Gate('cx', buildControlled(Gate.x));
+Gate.cy = new Gate('cy', buildControlled(Gate.y));
+Gate.cz = new Gate('cz', buildControlled(Gate.z));
+Gate.ch = new Gate('ch', buildControlled(Gate.h));
+Gate.csrn = new Gate('csrn', buildControlled(Gate.srn));
+Gate.cs = new Gate('cs', buildControlled(Gate.s));
+Gate.cr2 = new Gate('cr2', buildControlled(Gate.r2));
+Gate.cr4 = new Gate('cr4', buildControlled(Gate.r4));
+Gate.cr8 = new Gate('cr8', buildControlled(Gate.r8));
+
+const gates = new Map(Object.entries(Gate));
+
+module.exports = {
+  Gate,
+  gates
 };
-
-gates.cx = buildControlled(gates.x);
-gates.cy = buildControlled(gates.y);
-gates.cz = buildControlled(gates.z);
-gates.ch = buildControlled(gates.h);
-gates.csrn = buildControlled(gates.srn);
-gates.cs = buildControlled(gates.s);
-gates.cr2 = buildControlled(gates.r2);
-gates.cr4 = buildControlled(gates.r4);
-gates.cr8 = buildControlled(gates.r8);
-
-module.exports = gates;