+++ packages/qiskit-sim/test/functional/gates.js
@@ -15,10 +15,11 @@ const { gates } = require('../..');
 
 describe('sim:gates', () => {
   it('should include all supported ones', () =>
-    assert.deepEqual(Object.keys(gates), [
+    assert.deepEqual([...gates.keys()], [
       'x',
       'y',
       'z',
+      'id',
       'h',
       'srn',
       's',